Implemented the remove_merge_commits() function.
This function identifies and removes all merge commits that are no
longer needed after the removal of certain files from all commits in
a certain branch. This functionality was previously implemented as part
of the remove_files() function, but since this procedure is potentially
very aggressive, I decided to make it optional.